refactor: pass a list of activity_types
Not super happy with this kind of implementation… But this way most code can be reused.
For `deleted` the new email-template will be used, all other types will still be treated the same.
If we wanted to have a specific template for the other activity_types, too, this could be done.